QPPA-9422: refactor mvp helper functions
