+++ index.ts
@@ -2,8 +2,9 @@
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import * as YAML from 'yaml';
-import * as _ from 'lodash';
 import { Constants } from './constants';
+import { createMVPDataFile } from './mvp-data-utils';
+
 
 const yearRegEx = /^[0-9]{4}/;
 const benchmarkJsonFileRegEx = /^[0-9]{4}\.json$/;
@@ -175,7 +176,7 @@ export function getMVPData(performanceYear: number = 2023, mvpIds: string[] = []
   if (fse.existsSync(filePath)) {
     mvpData = JSON.parse(fse.readFileSync(filePath, 'utf8'));
   } else {
-    mvpData = exports.createMVPDataFile(performanceYear);
+    mvpData = createMVPDataFile(performanceYear);
   }
 
   if (mvpIds.length) {
@@ -183,82 +184,6 @@ export function getMVPData(performanceYear: number = 2023, mvpIds: string[] = []
   }
 
   return mvpData;
-}
-
-/**
- * @return {any}
- */
-export function createMVPDataFile(performanceYear: number): any {
-  const mvpFilePath = path.join(__dirname, 'mvp', performanceYear.toString(), 'mvp-enriched.json');
-  const measureFilePath = path.join(__dirname, 'measures', performanceYear.toString(), 'measures-data.json');
-
-  let mvpData: any[] = [];
-  let measuresData: any[] = [];
-  try {
-    mvpData = JSON.parse(
-        fse.readFileSync(path.join(__dirname, 'mvp', performanceYear.toString(), 'mvp.json'), 'utf8'));
-    measuresData = exports.getMeasuresData(performanceYear);
-  } catch (e) {
-    console.log('QPP mvp / measures data not found for year: ' + performanceYear + ' --> ' + e);
-    return [];
-  }
-
-  const mvpIds: any[] = [];
-  mvpData.forEach(mvpDataItem => {
-    if (mvpDataItem.mvpId !== 'app1') {
-      mvpIds.push(mvpDataItem.mvpId);
-    }
-  });
-
-  // Reset the allowedPrograms to remove any MVP ID program names (we will add them back later in the process)
-  // This allows for removal of a measure from an MVP data item
-  measuresData.forEach(measure => {
-    measure.allowedPrograms = measure.allowedPrograms ? measure.allowedPrograms.filter(program => !mvpIds.includes(program)) : [];
-  });
-
-  // Hydrate measures
-  mvpData.forEach(mvp => {
-    Constants.mvpMeasuresHelper.forEach(item => {
-      mvp[item.enrichedMeasureKey] = [];
-      exports.populateMeasuresforMVPs(mvp, mvpData, measuresData, item.measureIdKey, item.enrichedMeasureKey);
-    });
-
-    mvp.hasOutcomeAdminClaims = !_.isEmpty(mvp.administrativeClaimsMeasureIds);
-  });
-
-  mvpData.forEach(mvp => {
-    Constants.mvpMeasuresHelper.forEach(item => {
-      delete mvp[item.measureIdKey];
-    });
-  });
-
-  fse.writeFileSync(mvpFilePath, JSON.stringify(mvpData, null, 2));
-  fse.writeFileSync(measureFilePath, JSON.stringify(measuresData, null, 2));
-
-  return mvpData;
-}
-
-export function populateMeasuresforMVPs(currentMvp: any, allMvps: any[], measuresData: any[], measureIdKey: string, enrichedMeasureKey: string): void {
-  currentMvp[measureIdKey].forEach((measureId: string) => {
-    const measure = measuresData.find(measure => measure.measureId === measureId);
-
-    if (measure) {
-      allMvps.forEach(mvp => {
-        // update measuresData with MVP programNames.
-        if (mvp[measureIdKey].includes(measureId)) {
-          measure.allowedPrograms.push(mvp.mvpId);
-        }
-      });
-      measure.allowedPrograms = _.uniq(measure.allowedPrograms);
-
-      if (measure.measureId === '321') {
-        currentMvp.hasCahps = true;
-      }
-
-      // update mvp-data with measures.
-      currentMvp[enrichedMeasureKey].push(measure);
-    }
-  });
 }
 
 /**
